+++ packages/wallet-sdk/src/CoinbaseWalletProvider.ts
@@ -15,8 +15,8 @@ import { areAddressArraysEqual, prepend0x, showDeprecationWarning } from './core
 import { FilterRequestHandler } from './filter/FilterRequestHandler';
 import { StateRequestHandler } from './internalState/StateRequestHandler';
 import { RPCFetchRequestHandler } from './rpcFetch/RPCFetchRequestHandler';
+import { AccountsUpdate, ChainUpdate } from './sign/interface';
 import { SignRequestHandler } from './sign/SignRequestHandler';
-import { AccountsUpdate, ChainUpdate } from './sign/UpdateListenerInterface';
 
 export class CoinbaseWalletProvider extends EventEmitter implements ProviderInterface {
   protected accounts: AddressString[] = [];
