+++ packages/wallet-sdk/src/sign/SignRequestHandler.ts
@@ -1,6 +1,6 @@
+import { Signer } from './interface';
+import { SignRequestHandlerListener } from './interface';
 import { SignerConfigurator } from './SignerConfigurator';
-import { Signer } from './SignerInterface';
-import { SignRequestHandlerListener } from './UpdateListenerInterface';
 import { WLSigner } from './walletlink/WLSigner';
 import { PopUpCommunicator } from ':core/communicator/PopUpCommunicator';
 import { CB_KEYS_URL } from ':core/constants';
