+++ packages/wallet-sdk/src/sign/interface.ts
@@ -1,4 +1,11 @@
 import { AddressString, Chain } from ':core/type';
+import { RequestArguments } from ':core/type/ProviderInterface';
+
+export interface Signer {
+  handshake(): Promise<AddressString[]>;
+  request<T>(request: RequestArguments): Promise<T>;
+  disconnect: () => Promise<void>;
+}
 
 type UpdateSource = 'wallet' | 'storage';
