+++ packages/wallet-sdk/src/sign/scw/SCWSigner.ts
@@ -1,5 +1,5 @@
-import { Signer } from '../SignerInterface';
-import { StateUpdateListener } from '../UpdateListenerInterface';
+import { Signer } from '../interface';
+import { StateUpdateListener } from '../interface';
 import { SCWKeyManager } from './SCWKeyManager';
 import { SCWStateManager } from './SCWStateManager';
 import { PopUpCommunicator } from ':core/communicator/PopUpCommunicator';
